requirements.txt
PyAudioWPatch
pyserial